requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pypdf
google-generativeai
slowapi
python-json-logger
python-multipart
fpdf
pytest
chromadb
faiss-cpu
sentence-transformers